authserver,oidcauth: remove ConfigureOIDC hook indirection
Now that oidcauth lives in pkg/security/oidcauth (not under pkg/ccl),
the server package can import it directly. This removes the
ConfigureOIDC hook variable, the OIDC interface wrapper, and the
noOIDCConfigured stub type from authserver.

ConfigureOIDC is changed from a package-level var to a regular
function, and its return type simplified from authserver.OIDC to
ui.OIDCUI (the wrapper interface added nothing).

The server.ServerHTTPBasePath cluster setting is moved from
pkg/server to pkg/server/authserver to break the import cycle
that previously prevented oidcauth from being imported by server
(oidcauth needed server for this setting, creating oidcauth →
server → oidcauth).

The blank import of oidcauth is removed from ccl_init.go.

Release note: None
