+++ pkg/ccl/BUILD.bazel
@@ -17,6 +17,5 @@ go_library(
         "//pkg/crosscluster/logical",
         "//pkg/crosscluster/physical",
         "//pkg/crosscluster/producer",
-        "//pkg/security/oidcauth",
     ],
 )
