+++ pkg/security/oidcauth/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//pkg/security/jwthelper",
         "//pkg/security/provisioning",
         "//pkg/security/username",
-        "//pkg/server",
         "//pkg/server/authserver",
         "//pkg/server/serverpb",
         "//pkg/server/telemetry",
