+++ pkg/server/BUILD.bazel
@@ -182,6 +182,7 @@ go_library(
         "//pkg/security",
         "//pkg/security/certnames",
         "//pkg/security/clientsecopts",
+        "//pkg/security/oidcauth",
         "//pkg/security/username",
         "//pkg/server/apiconstants",
         "//pkg/server/apiinternal",
