+++ pkg/server/authserver/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
     deps = [
         "//pkg/base",
         "//pkg/multitenant",
-        "//pkg/roachpb",
         "//pkg/security",
         "//pkg/security/jwtauth",
         "//pkg/security/ldapauth",
@@ -37,12 +36,10 @@ go_library(
         "//pkg/sql/sessiondata",
         "//pkg/sql/syntheticprivilege",
         "//pkg/sql/types",
-        "//pkg/ui",
         "//pkg/util/grpcutil",
         "//pkg/util/log",
         "//pkg/util/protoutil",
         "//pkg/util/timeutil",
-        "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_logtags//:logtags",
         "@com_github_cockroachdb_redact//:redact",
